+++ schema/TABLE/metadata.sql
@@ -1,0 +1,4 @@
+CREATE TABLE IF NOT EXISTS metadata (
+    name text,
+    value text
+);