+++ schema/TABLE/tiles.sql
@@ -1,0 +1,6 @@
+CREATE TABLE IF NOT EXISTS tiles (
+    zoom_level integer,
+    tile_column integer,
+    tile_row integer,
+    tile_data blob
+);