+++ utils/utils.js
@@ -1,7 +1,8 @@
 const fs = require('fs')
 const mkdirp = require('mkdirp')
 const path = require('path')
-const Sequelize = require('sequelize-offline')
+const sqlite3 = require('sqlite3-offline')
+const chalk = require('chalk')
 
 /**
  * Create Folder
@@ -9,7 +10,7 @@ const Sequelize = require('sequelize-offline')
  * @param {string} uri
  * @returns {Promise<boolean>}
  */
-module.exports.createFolder = function (uri) {
+module.exports.createFolder = (uri) => {
   const dirname = path.dirname(uri)
   return new Promise((resolve, reject) => {
     if (!fs.existsSync(dirname)) {
@@ -17,6 +18,20 @@ module.exports.createFolder = function (uri) {
     }
     return resolve(true)
   })
+}
+
+/**
+ * Pretty Error message
+ */
+module.exports.error = (message) => {
+  throw new Error(chalk.bgRed.white(message))
+}
+
+/**
+ * Pretty Warning message
+ */
+module.exports.warning = (message) => {
+  console.log(chalk.bgYellow.black(message))
 }
 
 /**
@@ -28,7 +43,7 @@ module.exports.createFolder = function (uri) {
  * getFiles('/home/myfiles')
  * //=['map', 'test']
  */
-module.exports.getFiles = function (path, regex = /\.mbtiles$/) {
+module.exports.getFiles = (path, regex = /\.mbtiles$/) => {
   let mbtiles = fs.readdirSync(path).filter(value => value.match(regex))
   mbtiles = mbtiles.map(data => data.replace(regex, ''))
   mbtiles = mbtiles.filter(name => !name.match(/^_.*/))
@@ -39,16 +54,10 @@ module.exports.getFiles = function (path, regex = /\.mbtiles$/) {
  * Connect to SQL MBTiles DB
  *
  * @param {string} uri
- * @returns {Sequelize} Sequelize connection
+ * @returns {Sqlite3} Sqlite3 connection
  */
-module.exports.connect = function (uri) {
-  const options = {
-    define: { freezeTableName: true, timestamps: false },
-    logging: false,
-    pool: { idle: 10000, max: 5, min: 0 },
-    storage: uri
-  }
-  return new Sequelize('sqlite://' + uri, options)
+module.exports.connect = (uri) => {
+  return new sqlite3.Database(uri)
 }
 
 /**
@@ -56,68 +65,70 @@ module.exports.connect = function (uri) {
  * @param {ParseMetadata[]} data
  * @returns Metadata
  */
-module.exports.parseMetadata = function (data) {
+module.exports.parseMetadata = (data) => {
   const metadata = {}
-  data.map(item => {
-    const name = item.name.toLowerCase()
-    const value = item.value
-    switch (name) {
-      case 'minzoom':
-      case 'maxzoom':
-        metadata[name] = Number(value)
-        break
-      case 'name':
-      case 'attribution':
-      case 'description':
-        metadata[name] = value
-        break
-      case 'bounds':
-        const bounds = value.split(',').map(i => Number(i))
-        metadata.bounds = [bounds[0], bounds[1], bounds[2], bounds[3]]
-        break
-      case 'center':
-        const center = value.split(',').map(i => Number(i))
-        switch (center.length) {
-          case 2:
-            metadata.center = [center[0], center[1]]
-            break
-          case 3:
-            metadata.center = [center[0], center[1], center[2]]
-            break
-          default:
-        }
-        break
-      case 'type':
-        switch (value) {
-          case 'overlay':
-          case 'baselayer':
-            metadata[name] = value
-            break
-          default:
-        }
-        break
-      case 'format':
-        switch (value) {
-          case 'png':
-          case 'jpg':
-            metadata[name] = value
-            break
-          default:
-        }
-        break
-      case 'version':
-        switch (value) {
-          case '1.0.0':
-          case '1.1.0':
-          case '1.2.0':
-            metadata[name] = value
-            break
-          default:
-        }
-        break
-      default:
-        metadata[name] = value
-    }
-  })
+  if (data) {
+    data.map(item => {
+      const name = item.name.toLowerCase()
+      const value = item.value
+      switch (name) {
+        case 'minzoom':
+        case 'maxzoom':
+          metadata[name] = Number(value)
+          break
+        case 'name':
+        case 'attribution':
+        case 'description':
+          metadata[name] = value
+          break
+        case 'bounds':
+          const bounds = value.split(',').map(i => Number(i))
+          metadata.bounds = [bounds[0], bounds[1], bounds[2], bounds[3]]
+          break
+        case 'center':
+          const center = value.split(',').map(i => Number(i))
+          switch (center.length) {
+            case 2:
+              metadata.center = [center[0], center[1]]
+              break
+            case 3:
+              metadata.center = [center[0], center[1], center[2]]
+              break
+            default:
+          }
+          break
+        case 'type':
+          switch (value) {
+            case 'overlay':
+            case 'baselayer':
+              metadata[name] = value
+              break
+            default:
+          }
+          break
+        case 'format':
+          switch (value) {
+            case 'png':
+            case 'jpg':
+              metadata[name] = value
+              break
+            default:
+          }
+          break
+        case 'version':
+          switch (value) {
+            case '1.0.0':
+            case '1.1.0':
+            case '1.2.0':
+              metadata[name] = value
+              break
+            default:
+          }
+          break
+        default:
+          metadata[name] = value
+      }
+    })
+  }
   return metadata
 }